requirements.txt
asgiref==3.7.2
Django==5.0.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
djangorestframework==3.14.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
tzdata==2023.4
